REFACTOR: 컴포넌트 파일 분리 및 파라미터 수정
